Fix Algolia results for top-level pages
Previously, pressing the button on the top-level of popular domains
would return one of the top results for any path on that domain. This
fixes that so the result you're redirected to is the top result for the
actual current page, even if it's a domain common to other results on
different paths/pages.