Fix concurrency bug and add comments
Since storing a Bloom filter in local storage is not instant, clicking
too many links on news.ycombinator.com too fast would result in a second
call to storeBloom while the first was in the process of storing and had
bloom.addr set to null. This would typically end up causing the
window.bloom.addr value to be set to null, and would possibly store a
fucked up version of the Bloom filter. The bug has been fixed by adding
a semaphore.