Fix glitch in sendLatest
If one of the stories on the Hacker News front page did not have points,
then the length of stories and scores would differ, and an exception
would be thrown. This fixes that by using the parent node of score,
which will always exist, whether its child `.score` does or not.